functions/upload-world-file/requirements.txt
google-cloud-storage==1.10.0
jsonschema==2.6.0
flask-cors==3.0.6
